+++ consensus/state_sync/module.go
@@ -58,7 +58,7 @@ type stateSync struct {
 	//currentMode SyncMode
 	serverMode bool
 
-	logger    modules.Logger
+	logger    *modules.Logger
 	logPrefix string
 
 	// m is a mutex used to control synchronization when multiple goroutines are accessing the struct and its fields / properties.
@@ -150,7 +150,7 @@ func (m *stateSync) HandleGetBlockResponse(blockRes *typesCons.GetBlockResponse)
 		"receiver":      clientPeerId,
 	}
 
-	m.logger.Info().Fields(fields).Msgf("Received StateSync GetBlockResponse: %s", blockRes)
+	m.logger.Info().Fields(fields).Msgf("Received GetBlockResponse: %s", blockRes)
 
 	return nil
 }
@@ -169,7 +169,7 @@ func (m *stateSync) HandleStateSyncMetadataResponse(metaDataRes *typesCons.State
 		"receiver":      clientPeerId,
 	}
 
-	m.logger.Info().Fields(fields).Msgf("Received StateSync MetadataResponse: %s", metaDataRes)
+	m.logger.Info().Fields(fields).Msgf("Received StateSyncMetadataResponse: %s", metaDataRes)
 
 	return nil
 }