+++ shared/CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.34] - 2023-02-24
+
+- Remove SetLeaderId() method from ConsensusDebugModule interface
+
+## [0.0.0.33] - 2023-02-24
+
+- Update logger value references with pointers
+
+## [0.0.0.32] - 2023-02-23
+
+- Add file utility functions for checking, reading and writing to files
+
+## [0.0.0.31] - 2023-02-22
+
+-  Export consensus module's ConsensusDebugModule interface.
+
 ## [0.0.0.30] - 2023-02-21
 
 - Rename ServiceNode Actor Type Name to Servicer