+++ state_machine/docs/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.2] - 2023-02-24
+
+- Update logger value references with pointers
+
 ## [0.0.0.1] - 2023-02-17
 
 - Introduced this `CHANGELOG.md` and  `README.md`